Move watchEpilogue from HumanPlayer to PlayerIO; delegate in HumanPlayer
watchEpilogue の ChronicleView フォーマット処理を ConsolePlayerIO に移動し、
HumanPlayer は io.watchEpilogue(chronicles) に委譲するだけになった。
これで HumanPlayer と WebPlayer が watchEpilogue を同じ形で IO に委譲する状態になった。
