+++ src/main/kotlin/werewolf/human/HumanPlayer.kt
@@ -64,14 +64,6 @@ class HumanPlayer(role: Role, override val name: String, private val io: PlayerI
     }
 
     override fun watchEpilogue(chronicles: List<ChronicleView>) {
-        io.sendMessage("ゲーム振り返り", chronicles.joinToString("\n") { it.formatForConsole() })
-    }
-}
-
-private fun ChronicleView.formatForConsole(): String = when (this) {
-    is ChronicleView.Observation -> "[$recipient] [$category] $content"
-    is ChronicleView.Action -> {
-        val line = "[$actor] [$category] $content"
-        if (intent.isNotEmpty()) "$line\n  [$intent]" else line
+        io.watchEpilogue(chronicles)
     }
 }
