+++ src/main/kotlin/werewolf/human/PlayerIO.kt
@@ -1,5 +1,6 @@
 package werewolf.human
 
+import werewolf.game.ChronicleView
 import werewolf.game.RecallView
 import werewolf.view.ChoiceView
 
@@ -8,6 +9,7 @@ abstract class PlayerIO {
     abstract fun sendMessage(title: String, content: String)
     protected abstract fun readInput(): String
     abstract fun promptChoice(view: ChoiceView): String
+    abstract fun watchEpilogue(chronicles: List<ChronicleView>)
 
     fun promptFreeText(title: String, content: String): String {
         sendMessage(title, content)
